requirements.txt
boto3~=1.29.6
python-dotenv~=1.0.0
numpy~=1.24.4
flask~=3.0.0
waitress~=2.1.2
requests~=2.31.0